refactor : 패키지 정리 및 static 처리
