+++ src/main/java/lotto/Constants/Constants.java
@@ -1,15 +1,13 @@
 package lotto.Constants;
 
 public final class Constants {
-
-    private Constants() {
-    }
-
     public static final int LOTTO_PRICE = 1000;
     public static final int MIN_RANDOM_NUMBER = 1;
     public static final int MAX_RANDOM_NUMBER = 45;
     public static final int LOTTO_NUMBER_COUNT = 6;
-
     public static final int ZERO = 0;
+
+    private Constants() {
+    }
 
 }
