+++ src/main/java/lotto/error/exception/LottoException.java
@@ -1,4 +1,6 @@
-package lotto.exception;
+package lotto.error.exception;
+
+import lotto.error.ErrorCode;
 
 public class LottoException extends IllegalArgumentException {
