+++ src/main/java/lotto/error/exception/PurchaseException.java
@@ -1,4 +1,6 @@
-package lotto.exception;
+package lotto.error.exception;
+
+import lotto.error.ErrorCode;
 
 public class PurchaseException extends IllegalArgumentException {
