+++ src/main/java/lotto/model/WinningLotto.java
@@ -2,11 +2,11 @@ package lotto.model;
 
 import static lotto.Constants.Constants.MAX_RANDOM_NUMBER;
 import static lotto.Constants.Constants.MIN_RANDOM_NUMBER;
-import static lotto.exception.ErrorCode.BONUS_NUMBER_DUPLICATE;
-import static lotto.exception.ErrorCode.BONUS_NUMBER_RANGE;
+import static lotto.error.ErrorCode.BONUS_NUMBER_DUPLICATE;
+import static lotto.error.ErrorCode.BONUS_NUMBER_RANGE;
 
 import java.util.List;
-import lotto.exception.BonusNumberException;
+import lotto.error.exception.BonusNumberException;
 
 public class WinningLotto {
 
@@ -19,18 +19,18 @@ public class WinningLotto {
         this.bonus = bonus;
     }
 
-    private static void validateBonusNumber(final Lotto winningLotto, final int bonus) {
+    private void validateBonusNumber(final Lotto winningLotto, final int bonus) {
         validateBonusNumberRange(bonus);
         validateBonusNumberDuplicate(winningLotto, bonus);
     }
 
-    private static void validateBonusNumberRange(final int number) {
+    private void validateBonusNumberRange(final int number) {
         if (number < MIN_RANDOM_NUMBER || number > MAX_RANDOM_NUMBER) {
             throw new BonusNumberException(BONUS_NUMBER_RANGE);
         }
     }
 
-    private static void validateBonusNumberDuplicate(final Lotto winningLotto, final int number) {
+    private void validateBonusNumberDuplicate(final Lotto winningLotto, final int number) {
         if (winningLotto.contains(number)) {
             throw new BonusNumberException(BONUS_NUMBER_DUPLICATE);
         }
